requirements.txt
Flask
Flask-RESTFull
Flask-JWT
uwsgi